+++ src/views/SignIn.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Login from '../components/Login';
+import Login from '../components/molecules/Login';
 import AuthForm from '../layout/AuthForm';
 
 const SignIn = ( { onLogin } ) => (
